package/requirements.txt
# AI 学术写作助手 - 打包依赖
# 包含后端所有依赖 + PyInstaller

# FastAPI 后端依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.10.0
httpx==0.27.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
redis==5.0.1
aioredis==2.0.1
sse-starlette==3.0.3
psycopg2-binary

# Word 格式化模块依赖
mistune>=3.0.0
python-docx>=1.1.0
lxml>=4.9.0

# 修复 PyInstaller pkg_resources 错误
jaraco-text
jaraco-classes
jaraco-context
jaraco-functools
platformdirs
packaging
more-itertools

# 打包工具
pyinstaller==6.3.0
